requirements.txt
gradio
requests
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-huggingface
langchain-groq
langchain-tavily
langchain-chroma
langgraph
huggingface_hub
chromadb
arxiv
pymupdf
wikipedia
pgvector
python-dotenv